backend/requirements.txt
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asyncpg==0.30.0
bcrypt==4.3.0
certifi==2025.6.15
cffi==1.17.1
click==8.2.1
cryptography==45.0.5
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.1.2
fastapi==0.115.14
fastapi-users==12.1.3
fastapi-users-db-sqlalchemy==7.0.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
makefun==1.16.0
Mako==1.3.10
MarkupSafe==3.0.2
passlib==1.7.4
pwdlib==0.2.1
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.8.0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.7
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
aiohttp